LPS-73352 Move to separate plugin class
